DR2-1348 Add S3 copy lambda
This currently gets the user credentials from SSM to create an S3 client
with access to the preservica S3 bucket.

It then streams from our staging bucket using its own credentials and
streams it directly to preservica using their credentials. It never
touches the disk.

This will be changed once we have a role to do a copy rather than
streaming through the lambda.